Update Child Service Controller and Repository

+++ src/main/java/senai/tcc/zupiapi/zupibackend/model/Child.java
@@ -16,7 +16,7 @@ public class Child {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     private String name;
     private LocalDate birthDate;
@@ -37,7 +37,7 @@ public class Child {
 
     public Child() {}
 
-    public Child(Integer id, String name, LocalDate birthDate, String schoolClass, String condition, User responsible) {
+    public Child(Long id, String name, LocalDate birthDate, String schoolClass, String condition, User responsible) {
         this.id = id;
         this.name = name;
         this.birthDate = birthDate;
@@ -46,11 +46,11 @@ public class Child {
         this.responsible = responsible;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
